Add MemoryCredentialStore and use in auth tests
Replace mocked out MongoDB stores with an in-memory store with the same
interface as the MongoDB credential store. This ensures that our
in-memory store is an adequate replacement for the MongoDB store and
will eventually allow us to remove MongoDB as a dependency. This is
part of an effort to remove external service dependencies (AWS S3 and
MongoDB) from the Dripbox server so that the project can be run more
easily.