Implement DiskS3Store.listObjectsV2
The interface is slightly different from pure S3 because fs does not
recursively list subdirectories which is nicer to deal with. Also it now
throws if the folder does not exist instead of returning an empty
folder. The logic could also be cleaned up on the middleware side to
make this function have to do less work to be compatible but for now it
works.